Restructuring to support local paths to additional code: - Move DataSource out of base/input_dataset.py and into dedicated module base/datasource.py - Expand DataSource to check if location is a directory or local repository - Rename AdditionalCode.source_repo to AdditionalCode.source and change type to DataSource - Make checkout_target an optional attribute to AdditionalCode (default None) - Wrap existing AdditionalCode.get() inside conditional check for 'url','repository' - Rename 'source' and 'source_repo' to 'location' in Case.from_blueprint, Case.persist, and cstar_blueprint_roms_marbl_example.yaml Still todo: - update docstrings - update example notebook
